requirements.txt
openpyxl==3.1.5
openpyxl==3.1.5
Pillow==11.2.1
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
rasa_sdk==3.12.0
Requests==2.32.3
